# packages/contracts/openapi/vinops.openapi.yaml
openapi: 3.1.0
info:
  title: VinOps API - Core Vertical Slice
  version: 1.1.0
  license:
    name: Proprietary - VinOps internal project artifact
    identifier: LicenseRef-Proprietary
  description: >-
    Executable contract for the VIN-MEGA-001 platform foundation and VIN-MEGA-002 Document Control vertical slice.
    Tenant, project, action and resource scope are derived and verified server-side; client identifiers never grant
    access. Current revision is an explicit context pointer changed only by an atomic authorized publication.
  x-vinops-cr-002:
    effective_date: '2026-07-30'
    access_token:
      default_ttl_minutes: 15
      storage: browser-memory-only
      cache_control: no-store
    refresh_credential:
      storage: hashed-server-representation
      rotation: every-refresh
      reuse_response: revoke-session-family-and-emit-security-audit-signal
      default_idle_ttl_days: 7
      default_absolute_ttl_days: 30
    csrf:
      required_for: cookie-authenticated-mutations
      token_binding: session-and-browser-memory
      origin_check: exact-allowlist-with-referer-fallback
    cors:
      origins: exact-environment-allowlist
      credentials_with_wildcard: forbidden
      production_localhost_or_null_default: forbidden
  x-vinops-response-invariants:
    correlation_id: Every response, including errors and 204 responses, returns X-Correlation-ID and any problem body repeats it as correlation_id.
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /projects/{projectId}/sync/batches:
    post:
      operationId: ingestOfflineBatch
      tags:
        - OfflineSync
      summary: Bulk ingest offline ledger operations
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncBatchCreate'
      responses:
        '200':
          description: OK
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncBatchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /projects/{projectId}/changes:
    get:
      operationId: getProjectChanges
      tags:
        - OfflineSync
      summary: Incremental sync change feed
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: OK
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeFeedResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token stored in memory only.
    refreshCookie:
      type: apiKey
      in: cookie
      name: vinops_refresh
      description: Opaque rotated credential. Cookie is Secure, HttpOnly, SameSite=Lax, host-only, Path=/api/v1/auth; idle TTL defaults to 7 days and absolute TTL to 30 days. Never log its value.
  parameters:
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
...
  headers:
    CorrelationId:
      description: End-to-end request correlation identifier returned on every response, including errors and 204 responses.
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Malformed request
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required or session revoked
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required:
        - code
        - message_key
        - correlation_id
        - retryable
      properties:
        code:
          type: string
          example: PERMISSION_DENIED
        message_key:
          type: string
          example: errors.permissionDenied
        details:
          type: object
          additionalProperties: true
        correlation_id:
          type: string
          format: uuid
        retryable:
          type: boolean
...
    SyncOperationInput:
      type: object
      required:
        - operation_id
        - entity_type
        - entity_temp_id
        - command
        - base_version
        - payload
        - client_created_at
      properties:
        operation_id:
          type: string
          format: uuid
        entity_type:
          type: string
        entity_temp_id:
          type: string
        command:
          type: string
        base_version:
          type: integer
        payload:
          type: object
        payload_hash:
          type: string
        client_created_at:
          type: string
          format: date-time
        dependency_ids:
          type: array
          items:
            type: string
    SyncBatchCreate:
      type: object
      required:
        - device_id
        - operations
      properties:
        device_id:
          type: string
          format: uuid
        operations:
          type: array
          items:
            $ref: '#/components/schemas/SyncOperationInput'
    SyncOperationResult:
      type: object
      required:
        - operation_id
        - status
      properties:
        operation_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [applied, conflict, rejected, failed]
        canonical_id:
          type: string
          format: uuid
        canonical_version:
          type: integer
        error_code:
          type: string
        conflict_details:
          type: object
    SyncBatchResponse:
      type: object
      required:
        - batch_id
        - status
        - operations
      properties:
        batch_id:
          type: string
          format: uuid
        status:
          type: string
        operation_count:
          type: integer
        applied_count:
          type: integer
        conflict_count:
          type: integer
        operations:
          type: array
          items:
            $ref: '#/components/schemas/SyncOperationResult'
    ChangeFeedResponse:
      type: object
      required:
        - changes
      properties:
        changes:
          type: array
          items:
            type: object
            required:
              - entity_type
              - entity_id
              - version
              - action
            properties:
              entity_type:
                type: string
              entity_id:
                type: string
                format: uuid
              version:
                type: integer
              action:
                type: string
              data:
                type: object
        next_cursor:
          type: string
...